requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0

# Google Sheets Integration
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
